+++ lib/anomaly.rb
@@ -1,2 +1,7 @@
-require_relative "anomaly/version"
+# modules
 require_relative "anomaly/detector"
+require_relative "anomaly/version"
+
+module Anomaly
+  class Error < StandardError; end
+end